+++ src/main/java/roomescape/controller/dto/TimeResponses.java
@@ -1,7 +1,7 @@
 package roomescape.controller.dto;
 
 import java.util.List;
-import roomescape.domain.TimeSlot;
+import roomescape.domain.timeslot.TimeSlot;
 import roomescape.service.dto.AvailableTimeSlot;
 
 public record TimeResponses(
