+++ src/main/java/roomescape/domain/payment/PaymentConfirmation.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.payment;
 
 public record PaymentConfirmation(
         String paymentKey,
