+++ src/main/java/roomescape/domain/payment/PaymentGateway.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.payment;
 
 public interface PaymentGateway {
