+++ src/main/java/roomescape/domain/payment/PaymentResult.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.payment;
 
 public record PaymentResult(
         String paymentKey,
