+++ src/main/java/roomescape/domain/payment/PaymentStatus.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.payment;
 
 import java.util.Arrays;
