+++ src/main/java/roomescape/domain/reservation/ReservationAndWaiting.java
@@ -1,6 +1,8 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import java.time.LocalDate;
+import roomescape.domain.theme.Theme;
+import roomescape.domain.timeslot.TimeSlot;
 
 public record ReservationAndWaiting(
         long id,
