+++ src/main/java/roomescape/domain/reservation/ReservationRepository.java
@@ -1,10 +1,8 @@
-package roomescape.repository;
+package roomescape.domain.reservation;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
-import org.springframework.stereotype.Repository;
-import roomescape.domain.Reservation;
 
 public interface ReservationRepository {
