+++ src/main/java/roomescape/domain/reservation/ReservationStatus.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 public enum ReservationStatus {
     PAYMENT_PENDING,
