+++ src/main/java/roomescape/domain/reservation/UserReservations.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import java.util.ArrayList;
 import java.util.Comparator;
