+++ src/main/java/roomescape/domain/reservation/WaitingWithNumber.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 public record WaitingWithNumber(
         Reservation waiting,
