+++ src/main/java/roomescape/infra/persistence/JdbcReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.infra.persistence;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -13,11 +13,12 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationSlot;
-import roomescape.domain.ReservationStatus;
-import roomescape.domain.Theme;
-import roomescape.domain.TimeSlot;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.ReservationSlot;
+import roomescape.domain.reservation.ReservationStatus;
+import roomescape.domain.theme.Theme;
+import roomescape.domain.timeslot.TimeSlot;
+import roomescape.domain.reservation.ReservationRepository;
 
 @Repository
 public class JdbcReservationRepository implements ReservationRepository {
