+++ src/main/java/roomescape/service/ThemeService.java
@@ -2,11 +2,11 @@ package roomescape.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.Theme;
+import roomescape.domain.theme.Theme;
 import roomescape.exception.NotFoundException;
 import roomescape.exception.ResourceInUseException;
-import roomescape.repository.ReservationRepository;
-import roomescape.repository.ThemeRepository;
+import roomescape.domain.reservation.ReservationRepository;
+import roomescape.domain.theme.ThemeRepository;
 
 import java.time.LocalDate;
 import java.util.List;
