+++ src/test/java/roomescape/domain/ThemeTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.theme.Theme;
 
 public class ThemeTest {
