+++ src/test/java/roomescape/repository/FakeReservationRepository.java
@@ -6,7 +6,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import roomescape.domain.Reservation;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.ReservationRepository;
 
 public class FakeReservationRepository implements ReservationRepository {
