+++ src/test/java/roomescape/repository/FakeThemeRepository.java
@@ -1,12 +1,13 @@
 package roomescape.repository;
 
-import roomescape.domain.Theme;
+import roomescape.domain.theme.Theme;
 
 import java.time.LocalDate;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import roomescape.domain.theme.ThemeRepository;
 
 public class FakeThemeRepository implements ThemeRepository {
