+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.Theme;
+import roomescape.domain.theme.Theme;
 import roomescape.repository.FakeReservationRepository;
 import roomescape.repository.FakeThemeRepository;
